Rollout::configure() method
This new method makes it possible to configure all attributes of a
feature in just one call.

This commits also fixes a bug in Feature::addUser() which made it
possible to add duplicate users.

Finally it adds a logic check to Feature::setPercentage() to make sure
that percentage is between 0 and 100.